bigmodel-finetune/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
itsdangerous==0.24
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.4.0
SQLAlchemy==1.3.6
Werkzeug==1.0.1
gevent==22.10.2
gunicorn==20.1.0
itsdangerous==0.24
esdk-obs-python==3.22.2
click==6.7
PyJWT==2.4.0
pytz==2022.7
click==6.7